+++ src/js/BezierPath.d.ts
@@ -470,18 +470,18 @@ export declare class BezierPath implements SVGSerializable {
      * @method fromJSON
      * @param {string} jsonString - The string with the JSON data.
      * @throw An error if the string is not JSON or does not contain a bezier path object.
-     * @instance
+     * @static
      * @memberof BezierPath
      * @return {BezierPath} The parsed bezier path instance.
      **/
-    fromJSON(jsonString: string): BezierPath;
+    static fromJSON(jsonString: string): BezierPath;
     /**
      * Create a BezierPath instance from the given array.
      *
      * @method fromArray
      * @param {Vertex[][]} arr - A two-dimensional array containing the bezier path vertices.
      * @throw An error if the array does not contain proper bezier path data.
-     * @instance
+     * @static
      * @memberof BezierPath
      * @return {BezierPath} The bezier path instance retrieved from the array data.
      **/