+++ src/ts/BezierPath.ts
@@ -1150,11 +1150,11 @@ export class BezierPath implements SVGSerializable {
      * @method fromJSON
      * @param {string} jsonString - The string with the JSON data.
      * @throw An error if the string is not JSON or does not contain a bezier path object.
-     * @instance
+     * @static
      * @memberof BezierPath
      * @return {BezierPath} The parsed bezier path instance.
      **/
-    fromJSON( jsonString:string ) : BezierPath {
+    static fromJSON( jsonString:string ) : BezierPath {
 	var obj:any = JSON.parse( jsonString );
 	return BezierPath.fromArray( obj );
     };
@@ -1166,7 +1166,7 @@ export class BezierPath implements SVGSerializable {
      * @method fromArray
      * @param {Vertex[][]} arr - A two-dimensional array containing the bezier path vertices.
      * @throw An error if the array does not contain proper bezier path data.
-     * @instance
+     * @static
      * @memberof BezierPath
      * @return {BezierPath} The bezier path instance retrieved from the array data.
      **/